requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-multipart
passlib[bcrypt]
pydantic
python-jose[cryptography]
email-validator
transformers
